# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Restaurant Platform API
  version: 1.0.0
  description: >
    REST API para el sistema de gestión de pedidos, reseñas y analítica de
    restaurantes sobre MongoDB Atlas (CC3089 Bases de Datos 2).


    **Recursos principales**: Users, Restaurants, Menu Items, Carts, Orders,
    Reviews, Files (GridFS), Analytics.
servers:
  - url: http://localhost:3000
    description: Servidor local de desarrollo
...
paths:
...
  /menu-items:
    post:
      summary: Crear ítem de menú
      tags: [MenuItems]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [restaurantId, name, price, category]
              properties:
                restaurantId:
                  type: string
                name:
                  type: string
                  example: Pizza Margherita
                description:
                  type: string
                price:
                  type: number
                  example: 89.00
                category:
                  type: string
                  example: Platos Principales
                allergens:
                  type: array
                  items:
                    type: string
                  example: [gluten, lácteos]
                tags:
                  type: array
                  items:
                    type: string
                preparationTimeMin:
                  type: integer
                  example: 20
      responses:
        '201':
          description: Ítem creado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuItem'

    get:
      summary: Listar ítems de menú (filtros, text search, sort, paginación)
      tags: [MenuItems]
      parameters:
        - in: query
          name: restaurantId
          schema:
            type: string
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: available
          schema:
            type: string
            enum: ['true', 'false']
        - in: query
          name: allergen
          schema:
            type: string
          description: Filtrar por alérgeno (multikey index).
        - in: query
          name: price_lte
          schema:
            type: number
          description: Precio máximo.
        - in: query
          name: q
          schema:
            type: string
          description: Búsqueda full-text ($text).
        - in: query
          name: sort
          schema:
            type: string
            enum: [salesCount, price]
            default: salesCount
        - $ref: '#/components/parameters/skip'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Lista de ítems de menú.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MenuItem'

    delete:
      summary: Eliminar ítems por restaurante y categoría (deleteMany)
      tags: [MenuItems]
      parameters:
        - in: query
          name: restaurantId
          required: true
          schema:
            type: string
        - in: query
          name: category
          schema:
            type: string
      responses:
        '200':
          description: Ítems eliminados.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResult'
        '400':
          description: restaurantId requerido.
  /menu-items/many:
    post:
      summary: Crear varios ítems a la vez (insertMany)
      tags: [MenuItems]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [restaurantId, items]
              properties:
                restaurantId:
                  type: string
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                      price:
                        type: number
                      category:
                        type: string
                      allergens:
                        type: array
                        items:
                          type: string
                      tags:
                        type: array
                        items:
                          type: string
      responses:
        '201':
          description: Ítems insertados.
          content:
            application/json:
              schema:
                type: object
                properties:
                  insertedCount:
                    type: integer
  /menu-items/bulk:
    post:
      summary: Operaciones mixtas en lote (bulkWrite ordered:false)
      description: >
        Soporta insertOne, updateOne, updateMany, deleteOne en un solo request.
      tags: [MenuItems]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [operations]
              properties:
                operations:
                  type: array
                  items:
                    type: object
                  description: Array de operaciones bulkWrite.
            example:
              operations:
                - insertOne:
                    document:
                      restaurantId: "65f1..."
                      name: "Nuevo Plato"
                      price: 50
                      category: "Entradas"
                - updateOne:
                    filter:
                      _id: "65f2..."
                    update:
                      $set:
                        price: 60
      responses:
        '200':
          description: Resultado del bulkWrite.
          content:
            application/json:
              schema:
                type: object
                properties:
                  insertedCount:
                    type: integer
                  modifiedCount:
                    type: integer
                  deletedCount:
                    type: integer
                  upsertedCount:
                    type: integer
  /menu-items/{id}:
    get:
      summary: Obtener ítem de menú por ID
      tags: [MenuItems]
      parameters:
        - $ref: '#/components/parameters/objectId'
      responses:
        '200':
          description: Detalle del ítem.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuItem'
        '404':
          description: Ítem no encontrado.

    patch:
      summary: Actualizar ítem de menú
      tags: [MenuItems]
      parameters:
        - $ref: '#/components/parameters/objectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                price:
                  type: number
                category:
                  type: string
                allergens:
                  type: array
                  items:
                    type: string
                preparationTimeMin:
                  type: integer
      responses:
        '200':
          description: Ítem actualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateResult'
        '404':
          description: Ítem no encontrado.

    delete:
      summary: Eliminar ítem de menú
      tags: [MenuItems]
      parameters:
        - $ref: '#/components/parameters/objectId'
      responses:
        '200':
          description: Ítem eliminado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResult'
        '404':
          description: Ítem no encontrado.
  /menu-items/{id}/availability:
    patch:
      summary: Cambiar disponibilidad + cascade a carritos (arrayFilters)
      description: >
        Actualiza available del ítem y propaga el cambio a todos los carritos
        que lo contienen usando updateMany + arrayFilters.
      tags: [MenuItems]
      parameters:
        - $ref: '#/components/parameters/objectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [available]
              properties:
                available:
                  type: boolean
      responses:
        '200':
          description: Disponibilidad actualizada.
          content:
            application/json:
              schema:
                type: object
                properties:
                  itemUpdated:
                    type: integer
                  affectedCarts:
                    type: integer
  /menu-items/restaurant/{restaurantId}/category-price:
    patch:
      summary: Ajustar precios por categoría (updateMany + $mul)
      tags: [MenuItems]
      parameters:
        - in: path
          name: restaurantId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                category:
                  type: string
                  description: Categoría a afectar (opcional, si omite aplica a todo).
                multiplier:
                  type: number
                  example: 1.1
                  description: Factor multiplicador (1.1 = +10%).
      responses:
        '200':
          description: Precios actualizados.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateResult'
...
  /carts/items:
    post:
      summary: Agregar ítem al carrito (upsert + $push + $setOnInsert)
      description: >
        Si el carrito no existe, se crea automáticamente (upsert).
        El ítem se valida contra menu_items y se agrega con embedded snapshot.
      tags: [Carts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, menuItemId]
              properties:
                userId:
                  type: string
                menuItemId:
                  type: string
                quantity:
                  type: integer
                  default: 1
      responses:
        '200':
          description: Carrito actualizado con el nuevo ítem.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '404':
          description: Ítem de menú no encontrado o no disponible.
  /carts/items/{menuItemId}:
    patch:
      summary: Actualizar cantidad de ítem en carrito (positional $ operator)
      tags: [Carts]
      parameters:
        - in: path
          name: menuItemId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, quantity]
              properties:
                userId:
                  type: string
                quantity:
                  type: integer
                  minimum: 1
      responses:
        '200':
          description: Carrito actualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '400':
          description: quantity debe ser >= 1.
        '404':
          description: Carrito o ítem no encontrado.

    delete:
      summary: Eliminar ítem del carrito ($pull)
      tags: [Carts]
      parameters:
        - in: path
          name: menuItemId
          required: true
          schema:
            type: string
        - in: query
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Ítem eliminado del carrito.
          content:
            application/json:
              schema:
                type: object
                properties:
                  removed:
                    type: integer
...
  /files:
    get:
      summary: Listar archivos subidos
      tags: [Files]
      parameters:
        - $ref: '#/components/parameters/skip'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Lista de archivos en GridFS.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    _id:
                      type: string
                    filename:
                      type: string
                    length:
                      type: integer
                    contentType:
                      type: string
                    uploadDate:
                      type: string
                      format: date-time
  /files/upload:
    post:
      summary: Subir imagen a GridFS
      description: >
        Sube una imagen y opcionalmente la vincula a un menu_item o restaurant
        actualizando imageFileId o logoFileId.
      tags: [Files]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image]
              properties:
                image:
                  type: string
                  format: binary
                menuItemId:
                  type: string
                  description: Vincular imagen a este menu_item.
                restaurantId:
                  type: string
                  description: Vincular logo a este restaurant.
      responses:
        '201':
          description: Archivo subido exitosamente.
          content:
            application/json:
              schema:
                type: object
                properties:
                  fileId:
                    type: string
                  filename:
                    type: string
                  size:
                    type: integer
                  url:
                    type: string
                    example: /files/65f1c5b8...
        '400':
          description: No se subió archivo o formato inválido.
  /files/{id}:
    get:
      summary: Descargar imagen de GridFS
      tags: [Files]
      parameters:
        - $ref: '#/components/parameters/objectId'
      responses:
        '200':
          description: Stream de la imagen.
          content:
            image/*:
              schema:
                type: string
                format: binary
        '404':
          description: Archivo no encontrado.

    delete:
      summary: Eliminar archivo de GridFS + desvincular
      tags: [Files]
      parameters:
        - $ref: '#/components/parameters/objectId'
      responses:
        '200':
          description: Archivo eliminado y desvinculado.
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
                  fileId:
                    type: string
...
components:
  parameters:
    objectId:
      in: path
      name: id
      required: true
      schema:
        type: string
      description: ObjectId de MongoDB (24 caracteres hex).
    skip:
      in: query
      name: skip
      schema:
        type: integer
        minimum: 0
        default: 0
    limit:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    MenuItem:
      type: object
      properties:
        _id:
          type: string
        restaurantId:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        allergens:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        available:
          type: boolean
        preparationTimeMin:
          type: integer
        imageFileId:
          type: string
          nullable: true
        salesCount:
          type: integer
    Cart:
      type: object
      properties:
        _id:
          type: string
        userId:
          type: string
        restaurantId:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              menuItemId:
                type: string
              name:
                type: string
              price:
                type: number
              quantity:
                type: integer
              subtotal:
                type: number
              available:
                type: boolean
        subtotal:
          type: number
        hasUnavailableItems:
          type: boolean
        expiresAt:
          type: string
          format: date-time
...
    UpdateResult:
      type: object
      properties:
        updated:
          type: integer
          example: 1
    DeleteResult:
      type: object
      properties:
        deleted:
          type: integer
          example: 1
...